Forgot to push editor, but I continued today
Before I continued I added imgui to the project and setup the basics of ImGui to be able to render with it.

Now I managed to create a simple scene manager and an object picker. But need to develop component editor and more things. Will do it as soon as I finish with final examns.

+++ Engine/Source/Core/Scene/ObjectManager.cpp
@@ -84,7 +84,7 @@ Engine::Space* Engine::ObjectManager::AddSpace()
 void Engine::ObjectManager::AddSpace(Space* space_)
 {
 	mSpaces.push_back(space_);
-	space_->SetSceneIdx(mSpaces.size() - 1);
+	space_->SetSceneIdx(static_cast<int>(mSpaces.size()) - 1);
 }
 
 /// -----------------------------------------------------------------
